[O11Y-218] add parquet.ColumnReader API to allow reading columns of typed values from parquet files (#72)
* add parquet.ColumnReader

* remove unused code + test column readers

* fix internal/bits checks of AVX-512 features

* add documentation

* improve documentation

* fix heading

* Update column_reader.go


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update README.md


* Update page.go


* PR feedback: improve readability

+++ column_buffer.go
@@ -455,10 +455,10 @@ func (col *repeatedColumnBuffer) Page() BufferedPage {
 			for i := int64(0); i < numValues; i++ {
 				var err error
 				if buffer, err = col.base.ReadRowAt(buffer[:0], int64(row.offset)+i); err != nil {
-					return newErrorPage(int16(col.Column()), "reordering rows of repeated column: %w", err)
+					return newErrorPage(col.Column(), "reordering rows of repeated column: %w", err)
 				}
 				if err = column.base.WriteRow(buffer); err != nil {
-					return newErrorPage(int16(col.Column()), "reordering rows of repeated column: %w", err)
+					return newErrorPage(col.Column(), "reordering rows of repeated column: %w", err)
 				}
 			}
 		}
